test(cli): organize and strengthen Textual coverage
